backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
faiss-cpu==1.7.4
scikit-learn==1.3.2
scipy==1.11.4
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
httpx==0.25.2
